pyproject.toml
[project]
name = "mixology-calculator"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = ["pyyaml"]

[project.scripts]
mix = "calculate:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = ["calculate.py"]
